+++ src/test/unit/java/cloud/metaapi/sdk/meta_api/TerminalStateTest.java
@@ -118,11 +118,14 @@ class TerminalStateTest {
   @Test
   void testReturnsSpecifications() {
     assertTrue(state.getSpecifications().isEmpty());
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.00001; }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ symbol = "GBPUSD"; }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ symbol = "AUDNZD"; }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.0001; }});
-    state.onSymbolSpecificationsRemoved(1, Lists.list("AUDNZD"));
+    state.onSymbolSpecificationsUpdated(1, Lists.list(
+      new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.00001; }},
+      new MetatraderSymbolSpecification() {{ symbol = "GBPUSD"; }}
+    ), Lists.list());
+    state.onSymbolSpecificationsUpdated(1, Lists.list(
+      new MetatraderSymbolSpecification() {{ symbol = "AUDNZD"; }},
+      new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.0001; }}
+    ), Lists.list("AUDNZD"));
     assertEquals(2, state.getSpecifications().size());
     assertThat(state.getSpecifications()).usingRecursiveComparison().isEqualTo(Lists.list(
       new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.0001; }},
@@ -185,16 +188,10 @@ class TerminalStateTest {
       profit = 100.0;
       volume = 2;
     }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ 
-      symbol = "EURUSD";
-      tickSize = 0.01;
-      digits = 5;
-    }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{ 
-      symbol = "AUDUSD";
-      tickSize = 0.01;
-      digits = 5;
-    }});
+    state.onSymbolSpecificationsUpdated(1, Arrays.asList(
+      new MetatraderSymbolSpecification() {{ symbol = "EURUSD"; tickSize = 0.01; digits = 5; }},
+      new MetatraderSymbolSpecification() {{ symbol = "AUDUSD"; tickSize = 0.01; digits = 5; }}
+    ), Arrays.asList());
     state.onSymbolPricesUpdated(1, Arrays.asList(new MetatraderSymbolPrice() {{
       time = new IsoTime();
       symbol = "EURUSD";
@@ -255,10 +252,10 @@ class TerminalStateTest {
       type = OrderType.ORDER_TYPE_SELL_LIMIT;
       currentPrice = 9;
     }});
-    state.onSymbolSpecificationUpdated(1, new MetatraderSymbolSpecification() {{
+    state.onSymbolSpecificationsUpdated(1, Arrays.asList(new MetatraderSymbolSpecification() {{
       symbol = "EURUSD";
       tickSize = 0.01;
-    }});
+    }}), Arrays.asList());
     state.onSymbolPricesUpdated(1, Arrays.asList(new MetatraderSymbolPrice() {{
       time = new IsoTime();
       symbol = "EURUSD";